organize injected files
