+++ src/injected-ui/timeline.svelte
